add data for training GANs, and add angle labels for images

+++ a3daps/synthetic_data.py
@@ -1,5 +1,6 @@
 from common.caching import read_input_dir, cached
 
+from . import dataio
 from collections import Counter
 
 import matplotlib.pyplot as plt
@@ -13,6 +14,9 @@ import skimage.io
 import skimage.transform
 import skimage.color
 import tqdm
+import random
+import itertools
+import skimage.transform
 
 
 def _convert_colors_to_label(image):
@@ -45,21 +49,22 @@ def _convert_colors_to_label(image):
     return ret
 
 
-@cached(version=4)
+@cached(version=8)
 def render_synthetic_body_zone_data(mode):
-    assert mode in ('sample', 'all', 'sample_train', 'train', 'sample_valid', 'valid')
+    assert mode in ('sample', 'sample_large', 'all')
 
     num_angles = 64
     image_size = 256
     with read_input_dir('makehuman/a3daps/meshes'):
         mesh_paths = sorted(['%s/%s' % (os.getcwd(), x) for x in glob.glob('*.mhx2')])
-    if mode.endswith('train'):
-        mesh_paths = mesh_paths[100:]
-    elif mode.endswith('valid'):
-        mesh_paths = meshes_paths[:100]
-    if mode.startswith('sample'):
+        random.seed(0)
+        random.shuffle(mesh_paths)
+    if mode == 'sample':
         mesh_paths = mesh_paths[:10]
-        num_angles = 4
+        num_angles = 8
+    elif mode == 'sample_large':
+        mesh_paths = mesh_paths[:100]
+
     done = Counter([x.split('_')[0] for x in glob.glob('*.png')])
     done = set([x for x, y in done.items() if y == 2*num_angles])
     todo = [x for x in mesh_paths if x.split('/')[-1].split('.')[0] not in done]
@@ -82,9 +87,10 @@ def render_synthetic_body_zone_data(mode):
     if not os.path.exists('done'):
         f = h5py.File('data.hdf5', 'w')
         x = f.create_dataset('x', (num_angles*len(mesh_paths), 2, image_size, image_size))
+        y = np.zeros(num_angles*len(mesh_paths), dtype='int32')
 
-        for i, file in enumerate(tqdm.tqdm(glob.glob('*_metal.png'))):
-            color_file = file.replace('metal', 'color')
+        for i, file in enumerate(tqdm.tqdm(glob.glob('*_skin.png'))):
+            color_file = file.replace('skin', 'color')
             image = skimage.color.rgb2gray(skimage.io.imread(file))
             image = skimage.transform.resize(image, (image_size, image_size))
             color = skimage.io.imread(color_file)
@@ -92,6 +98,50 @@ def render_synthetic_body_zone_data(mode):
             labels = _convert_colors_to_label(color[..., 0:3])
             x[i, 0, ...] = image
             x[i, 1, ...] = labels
+            y[i] = int(file.split('_')[1])
+
+        np.save('y.npy', y)
+        open('done', 'w').close()
+    else:
+        f = h5py.File('data.hdf5', 'r')
+        x = f['x']
+        y = np.load('y.npy')
+    return x, y
+
+
+@cached(render_synthetic_body_zone_data, version=0)
+def get_real_and_fake_images(mode):
+    if mode.startswith('sample'):
+        chunk_size = 128
+    else:
+        chunk_size = 1024
+    image_size = 256
+    num_angles = 64
+
+    if not os.path.exists('done'):
+        def get_real_images_chunk():
+            ret = []
+            while True:
+                for _, _, data in dataio.get_data(mode):
+                    ret.append(data)
+                    if len(ret) * num_angles >= chunk_size:
+                        ret = np.concatenate(ret, axis=-1)[..., :chunk_size]
+                        ret = skimage.transform.resize(ret, (image_size, image_size))
+                        ret = np.swapaxes(ret, 0, 2)[:, ::-1, :]
+                        yield ret
+                        ret = []
+
+        fake_images = render_synthetic_body_zone_data(mode)
+        f = h5py.File('data.hdf5', 'w')
+        x = f.create_dataset('x', fake_images.shape)
+
+        for i, real_chunk in zip(range(0, len(fake_images), chunk_size), get_real_images_chunk()):
+            rem = min(i+chunk_size, len(fake_images)) - i
+            real, fake = real_chunk[:rem], fake_images[i:i+rem, 0, ...]
+            np.random.shuffle(real)
+            np.random.shuffle(fake)
+            x[i:i+rem, 0, ...] = real
+            x[i:i+rem, 1, ...] = fake
 
         open('done', 'w').close()
     else:
